tests/wpt/web-platform-tests/offscreen-canvas/path-objects/2d.path.lineTo.nonfinite.details.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.path.lineTo.nonfinite.details</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.path.lineTo.nonfinite.details</h1>
<p class="desc">lineTo() with Infinity/NaN for first arg still converts the second arg</p>


<script>
var t = async_test("lineTo() with Infinity/NaN for first arg still converts the second arg");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

for (var arg1 of [Infinity, -Infinity, NaN]) {
  var converted = false;
  ctx.lineTo(arg1, { valueOf: function() { converted = true; return 0; } });
  _assert(converted, "converted");
}
t.done();

});
</script>
